Move PlotRef-related types to "types"; introduce guard for FacetedPlotRef
