+++ src/plots/PlotlyPlot.tsx
@@ -14,6 +14,7 @@ import React, {
 import { PlotParams } from 'react-plotly.js';
 import { legendSpecification } from '../utils/plotly';
 import Spinner from '../components/Spinner';
+import { PlotRef } from '../types/plots';
 import {
   PlotLegendAddon,
   PlotSpacingAddon,
@@ -26,15 +27,6 @@ import { select } from 'd3';
 import { ToImgopts, toImage } from 'plotly.js';
 import { uniqueId } from 'lodash';
 import { makeSharedPromise } from '../utils/promise-utils';
-
-/**
- * A generic imperative interface to plota. This allows us to create a facade
- * to interact with plot internals, such as exporting an image.
- */
-
-export interface PlotRef {
-  toImage: (imageOpts: ToImgopts) => Promise<string>;
-}
 
 export interface PlotProps<T> extends ColorPaletteAddon {
   /** plot data - following web-components' API, not Plotly's */
@@ -361,10 +353,6 @@ function PlotlyPlot<T>(
     </Suspense>
   );
 }
-
-export type ComponentWithPlotRef<P> = ComponentType<
-  PropsWithoutRef<P> & RefAttributes<PlotRef>
->;
 
 const PlotlyPlotWithRef = forwardRef(PlotlyPlot);
